+++ assets/src/scripts/components/well-log/index.spec.js
@@ -80,8 +80,7 @@ describe('well-log component', () => {
     let store;
 
     beforeEach(() => {
-        div = select('body')
-            .html(MOCK_COMPONENT);
+        div = select('body').append('div').html(MOCK_COMPONENT);
         store = getMockStore();
     });